+++ Study_SpringBoot/src/main/java/com/unoSpringBoot/study/Controller/Todo/TodoController.java
@@ -1,4 +1,4 @@
-package com.unoSpringBoot.study.Controller;
+package com.unoSpringBoot.study.Controller.Todo;
 
 import java.util.ArrayList;
 import java.util.List;
