+++ Study_SpringBoot/src/main/java/com/unoSpringBoot/study/Controller/Todo/TodoRetrieveTodoListCO.java
@@ -1,4 +1,4 @@
-package com.unoSpringBoot.study.Controller;
+package com.unoSpringBoot.study.Controller.Todo;
 
 import java.util.List;
